refactor home page component and replace relative imports in sass
