+++ next.config.js
@@ -1,10 +1,14 @@
 const createNextIntlPlugin = require('next-intl/plugin');
 const withNextIntl = createNextIntlPlugin();
+const path = require('path');
 
 /** @type {import('next').NextConfig} */
 
 const nextConfig = {
   reactStrictMode: true,
+  sassOptions: {
+    includePaths: [path.join(__dirname, 'src', 'app', 'styles')],
+  },
 };
 
 module.exports = withNextIntl(nextConfig);
