+++ src/app/hooks/useGame.ts
@@ -2,6 +2,7 @@ import { useContext } from 'react';
 
 import { GameContext } from '@/store/Game/GameProvider';
 
+// A simple hook to provide access to the GameContext.
 export const useGame = () => {
   const context = useContext(GameContext);
   if (!context) {
