+++ src/app/styles/pages/Home/components/Board.module.scss
@@ -1,4 +1,4 @@
-@import '../../../_mixins.module.scss';
+@import '_mixins.module.scss';
 
 $cell_width: (
   desktop: 160px,
