+++ src/app/styles/pages/Home/components/GameOptions.module.scss
@@ -1,4 +1,4 @@
-@import '../../../_mixins.module.scss';
+@import '_mixins.module.scss';
 
 $restartBtn_font_size: (
   desktop: 1.5rem,
