+++ src/app/styles/pages/Home/components/GameStatus.module.scss
@@ -1,4 +1,4 @@
-@import '../../../_mixins.module.scss';
+@import '_mixins.module.scss';
 
 $start_font_size: (
   desktop: 2rem,
